refactor: Move primitives nodes to individual files
Implements ML-597
